_tools/frontmatter_normalizer/requirements.txt
# Core dependencies
pyyaml>=6.0
click>=8.0

# NLP for category classification (SpaCy with large model for best vectors)
# Install model: python -m spacy download en_core_web_lg
# Or via URL: pip install https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
spacy>=3.0

# Fuzzy string matching for tags
rapidfuzz>=3.0

# Testing
pytest>=7.0
